replace uint8 by uint256 for loop and  array length

+++ contracts/interfaces/opty/IStrategyManager.sol
@@ -18,14 +18,14 @@ interface IStrategyManager {
      * @param _investStrategyHash Hash of the strategy being used in vault contract
      * @return Returns the withdrawal codes steps count for the given stretagy hash
      */
-    function getWithdrawAllStepsCount(bytes32 _investStrategyHash) external view returns (uint8);
+    function getWithdrawAllStepsCount(bytes32 _investStrategyHash) external view returns (uint256);
 
     /**
      * @dev Get the deposit codes steps count for the given stretagy hash
      * @param _investStrategyHash Hash of the strategy being used in vault contract
      * @return Returns the deposit codes steps count for the given stretagy hash
      */
-    function getDepositAllStepCount(bytes32 _investStrategyHash) external view returns (uint8);
+    function getDepositAllStepCount(bytes32 _investStrategyHash) external view returns (uint256);
 
     /**
      * @dev Get the claim reward token codes steps count for the given stretagy hash
